# 230-api/swagger.json
{ "openapi":"3.0.0",
  "info":{
    "title":"CAB230 - Web Computing API",
    "description":"This is Samuel Holden's CAB230 server side API for final assessment",
    "version":"1.0.0", "termsOfService":"http://swagger.io/terms/", "contact":{"email":"n10110381n@qut.edu.au"}, "license":{
      "name":"Apache 2.0",
      "url":"http://www.apache.org/licenses/LICENSE-2.0.html" }
  },
  "servers":[
    {"url": "https://localhost"},
    {"url": "https://172.22.24.250"}
  ],
  "paths": {
    "/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new user account",
        "description": "Registers a new account, for use of the secure search route",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "required": true,
                    "example": "user@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "required": true,
                    "example": "userPassword"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Account successfully created!"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error - User already exists"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login with an existing user account",
        "description": "Login to an existing account, in order to use the secure search route",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "required": true,
                    "example": "user@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "required": true,
                    "example": "userPassword"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6InRlc3QiLCJpYXQiOjE1NTg1MTc1MTgsImV4cCI6MTU1ODUxNzUzOH0.F2fjmSUq1p3e1kyOefm8tHSNoZDhND5KYm1HWJIytAE"
                    },
                    "token_type": {
                      "type": "string",
                      "example": "Bearer"
                    },
                    "expires_in": {
                      "type": "integer",
                      "example": 21600
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully logged in!"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User credentials do not match any of our records"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
  }
}